pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mystool"
version = "v1.1.0"
description = "NoneBot2插件|米游社工具-每日米游币任务、游戏签到、商品兑换、免抓包登录、原神树脂提醒"
license = "MIT"
authors = [
  "Ljzd-PRO <ljzd@office.ljzd-pro.ml>",
  "Everything0519 <598139245@qq.com>"
]
readme = "README.md"
homepage = "https://github.com/Ljzd-PRO/nonebot-plugin-mystool"
repository = "https://github.com/Ljzd-PRO/nonebot-plugin-mystool"
documentation = "https://github.com/Ljzd-PRO/nonebot-plugin-mystool/wiki"
keywords = ["bot", "qq", "qqbot", "onebotv11", "onebot", "nonebot2", "nonebot", "mihoyo", "mihoyobbs"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable"
]
packages = [
    { include = "nonebot_plugin_mystool", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
httpx = "^0.24.1"
nonebot_plugin_apscheduler = ">=0.2.0"
ntplib = "^0.4.0"
Pillow = ">=9.5,<11.0"
requests = "^2.31.0"
nonebot_adapter_onebot = "^2.2.3"
tenacity = "^8.2.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Ljzd-PRO/nonebot-plugin-mystool/issues"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
